requirements-dev.txt
# Development dependencies - Testing and code quality tools
# These are only needed for development, not production

# Include core dependencies for testing
-r requirements-core.txt

# Testing framework
pytest>=7.4.0,<8.0.0
pytest-cov>=4.1.0,<5.0.0
pytest-mock>=3.11.0,<4.0.0
pytest-asyncio>=0.21.0,<1.0.0
pytest-benchmark>=4.0.0,<5.0.0
pytest-timeout>=2.1.0,<3.0.0
pytest-xdist>=3.3.0,<4.0.0

# Advanced testing frameworks
hypothesis>=6.75.0,<7.0.0  # Property-based testing
pytest-randomly>=3.15.0,<4.0.0  # Random test ordering
pytest-repeat>=0.9.1,<1.0.0  # Test repetition for flaky tests
pytest-rerunfailures>=12.0,<13.0  # Rerun failed tests
pytest-html>=4.1.0,<5.0.0  # HTML test reports
pytest-json-report>=1.5.0,<2.0.0  # JSON test reports

# Load testing and performance
locust>=2.17.0,<3.0.0  # Load testing framework
pytest-benchmark>=4.0.0,<5.0.0  # Performance benchmarking
memory-profiler>=0.61.0,<1.0.0  # Memory profiling
fastapi>=0.100.0,<1.0.0  # FastAPI for mock server
uvicorn>=0.23.0,<1.0.0  # ASGI server for FastAPI

# Chaos engineering
chaostoolkit>=1.15.0,<2.0.0  # Chaos engineering toolkit
chaostoolkit-kubernetes>=0.26.0,<1.0.0  # K8s chaos experiments

# Mutation testing
mutmut>=2.4.0,<3.0.0
cosmic-ray>=8.4.0,<9.0.0

# Security testing
bandit>=1.7.5,<2.0.0
safety>=2.3.0,<3.0.0
semgrep>=1.40.0,<2.0.0
pip-audit>=2.6.0,<3.0.0

# Code quality and formatting
black>=23.7.0,<24.0.0
isort>=5.12.0,<6.0.0
ruff>=0.0.284,<1.0.0
mypy>=1.5.0,<2.0.0
pre-commit>=3.3.0,<4.0.0

# Development tools
jupyter>=1.0.0,<2.0.0
ipykernel>=6.25.0,<7.0.0
notebook>=7.0.0,<8.0.0

# Additional development utilities
click>=8.1.0,<9.0.0

# Type checking stubs
pandas-stubs>=2.3.0,<3.0.0

# Documentation testing
pydocstyle>=6.3.0,<7.0.0

# Sphinx documentation
docutils>=0.18.1,<0.20
sphinx>=7.0.0,<8.0.0
sphinx-rtd-theme>=1.3.0,<2.0.0
sphinx-autodoc-typehints>=1.24.0,<2.0.0
myst-parser>=2.0.0,<3.0.0
sphinx-autobuild>=2021.3.14

# Performance testing
pytest-profiling>=1.7.0,<2.0.0
memory-profiler>=0.61.0,<1.0.0

# Configuration validation
pydantic>=2.0.0,<3.0.0
jsonschema>=4.19.0,<5.0.0

# Data validation and testing
great-expectations>=0.17.0,<1.0.0  # Data quality testing
pandera>=0.17.0,<1.0.0  # Data validation

# Contract testing
pact-python>=1.0.0,<2.0.0  # Consumer-driven contract testing

# Test data generation
faker>=20.0.0,<21.0.0  # Fake data generation
factory-boy>=3.3.0,<4.0.0  # Test data factories
